# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /polymarket/compound:
    get:
      operationId: getPolymarketCompound
      tags: [polymarket]
      summary: Calculate the final stake from wallet and market liquidity
      parameters:
        - name: commonDepth
          in: query
          required: true
          description: Fresh common executable depth for the YES and NO asks, in shares
          schema:
            type: number
            minimum: 0
        - name: combinedAsk
          in: query
          required: true
          description: The current YES plus NO best ask, in pUSD per protected pair
          schema:
            type: number
            minimum: 0
            maximum: 2
      responses:
        "200":
          description: Current compound stake configuration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PolymarketCompound"
        "400":
          description: Missing or invalid market snapshot parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Invalid server balance configuration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /polymarket/market:
    get:
      operationId: getPolymarketMarket
      tags: [polymarket]
      summary: Read current public market data through the configured proxy
      parameters:
        - name: conditionId
          in: query
          required: true
          schema:
            type: string
            minLength: 1
      responses:
        "200":
          description: Market snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PolymarketMarket"
        "503":
          description: Proxy or upstream unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    PolymarketCompound:
      type: object
      properties:
        walletMaxStakePusd:
          type: number
        commonDepthShares:
          type: number
        combinedAskPusd:
          type: number
        marketAvailableVolumePusd:
          type: number
        finalExecutionStakePusd:
          type: number
        executable:
          type: boolean
        executionReason:
          type: string
          enum: [WALLET_LIMIT, MARKET_LIQUIDITY_LIMIT, NO_EXECUTABLE_VOLUME]
        balanceSource:
          type: string
          enum: [POLYMARKET_CLOB_COLLATERAL_BALANCE]
        marketSource:
          type: string
          enum: [REQUEST_MARKET_SNAPSHOT]
      required:
        - walletMaxStakePusd
        - commonDepthShares
        - combinedAskPusd
        - marketAvailableVolumePusd
        - finalExecutionStakePusd
        - executable
        - executionReason
        - balanceSource
        - marketSource
...
    PolymarketMarket:
      type: object
      properties:
        conditionId:
          type: string
        fetchedAt:
          type: string
          format: date-time
        raw:
          type: object
          additionalProperties: true
      required:
        - conditionId
        - fetchedAt
        - raw
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error